structure the code and removed redundant code
