+++ index.js
@@ -1,28 +1,5 @@
 const hello = () => {
     return;
 };
-const multiply = (a, b) => {
-    return a * b;
-};
-
-const add = (a, b) => {
-    return a + b;
-};
-
-const subtract = (a, b) => {
-    return a - b;
-};
-const divide = (a, b) => {
-    if (b === 0)
-        return undefined
-    return a / b;
-};
 
 hello();
-
-module.exports = {
-    add,
-    multiply,
-    subtract,
-    divide
-};
