+++ src/methods/add.js
@@ -1,6 +1,6 @@
 const add = (a, b) => {
     if (typeof a === "string" || typeof b === "string") {
-      return "Please Provide Numbers";
+      return "TypeError: Please Provide Numbers";
     }
     return a + b;
   };
