--- src/methods/substract.js
@@ -1,6 +1,0 @@
-const substract = (a, b) => {
-  if (typeof a === "string" || typeof b === "string") {
-    return "Please Provide Numbers";
-  }
-  return a - b;
-};
