+++ src/methods/subtract.js
@@ -1,0 +1,7 @@
+const subtract = (a, b) => {
+    if (typeof a === "string" || typeof b === "string") {
+        return "Please Provide Numbers";
+    }
+    return a - b;
+};
+module.exports = subtract;
